refactor(foundation): promote FitnessFunction/Scope/FunctionType from core (#602)
The three POD type definitions (`FitnessFunction`, `FitnessScope`,
`FitnessFunctionType`) moved from `mockforge-core::contract_drift::fitness`
into `mockforge-foundation::contract_drift_types` alongside the existing
`FitnessTestResult` (which was already there).

Why: the same three types were duplicated in
`mockforge-contracts::contract_drift::fitness` — defined as separate
Rust structs in two places. Promoting them to foundation breaks the
duplication and matches the pattern used for the other shared
contract-drift types (`AppType`, `ConsumerImpact`, `ConsumerMapping`,
`FitnessTestResult`, `SemanticDriftResult`, etc.).

Core's `fitness.rs` keeps its `FitnessFunctionRegistry` evaluator logic;
it now sources the three types from foundation via `pub use`. The
contracts duplicate of `fitness.rs` is deleted in the next commit.

No new derives added. Foundation's existing types from `FitnessTestResult`
onward do not gate on the `schema` feature, and core's existing
`FitnessFunction*` types didn't either; the promoted definitions preserve
that convention.
